News test and control of gameobjects

+++ src/3_domain/core/alis.core/Vector3D.cs
@@ -6,21 +6,21 @@ namespace Alis.Core
 {
     public struct Vector3D
     {
-        private int x;
+        private float x;
 
-        private int y;
+        private float y;
 
-        private int z;
+        private float z;
 
-        public Vector3D(int x, int y, int z)
+        public Vector3D(float x, float y, float z)
         {
             this.x = x;
             this.y = y;
             this.z = z;
         }
 
-        public int X { get => x; set => x = value; }
-        public int Y { get => y; set => y = value; }
-        public int Z { get => z; set => z = value; }
+        public float X { get => x; set => x = value; }
+        public float Y { get => y; set => y = value; }
+        public float Z { get => z; set => z = value; }
     }
 }
